+++ bflibrary/src/general/gline.c
@@ -21,29 +21,29 @@
 #include "bfline.h"
 #include "bfscreen.h"
 
-static void draw_b_line(long x1, long y1, long x2, long y2, TbPixel colour)
+static void draw_b_line(i32 x1, i32 y1, i32 x2, i32 y2, TbPixel colour)
 {
-  long apx = 2 * abs(x2 - x1);
-  long spx;
+  i32 apx = 2 * abs(x2 - x1);
+  i32 spx;
   if (x2 - x1 <= 0)
     spx = -1;
   else
     spx = 1;
-  long apy = 2 * abs(y2 - y1);
-  long spy;
+  i32 apy = 2 * abs(y2 - y1);
+  i32 spy;
   if (y2 - y1 <= 0 )
     spy = -1;
   else
     spy = 1;
-  long doffy = spy * lbDisplay.GraphicsScreenWidth;
-  long offset = lbDisplay.GraphicsScreenWidth * y1 + x1;
-  long x = x1;
-  long y = y1;
+  i32 doffy = spy * lbDisplay.GraphicsScreenWidth;
+  i32 offset = lbDisplay.GraphicsScreenWidth * y1 + x1;
+  i32 x = x1;
+  i32 y = y1;
   if (lbDisplay.DrawFlags & Lb_SPRITE_TRANSPAR4)
   {
     if (apx <= apy)
     {
-      long d = apx - (apy>>1);
+      i32 d = apx - (apy>>1);
       while ( true )
       {
         unsigned short glass_idx = lbDisplay.GraphicsWindowPtr[offset]
@@ -61,7 +61,7 @@ static void draw_b_line(long x1, long y1, long x2, long y2, TbPixel colour)
       }
     } else
     {
-      long d = apy - (apx >> 1);
+      i32 d = apy - (apx >> 1);
       while ( true )
       {
         unsigned short glass_idx = lbDisplay.GraphicsWindowPtr[offset]
@@ -83,7 +83,7 @@ static void draw_b_line(long x1, long y1, long x2, long y2, TbPixel colour)
   {
       if ( apx <= apy )
       {
-        long d = apx - (apy >> 1);
+        i32 d = apx - (apy >> 1);
         while ( 1 )
         {
           unsigned short glass_idx = (lbDisplay.GraphicsWindowPtr[offset]<<8)
@@ -101,7 +101,7 @@ static void draw_b_line(long x1, long y1, long x2, long y2, TbPixel colour)
         }
       } else
       {
-        long d = apy - (apx >> 1);
+        i32 d = apy - (apx >> 1);
         while ( 1 )
         {
           unsigned short glass_idx = (lbDisplay.GraphicsWindowPtr[offset]<<8)
@@ -122,7 +122,7 @@ static void draw_b_line(long x1, long y1, long x2, long y2, TbPixel colour)
   {
       if ( apx <= apy )
       {
-        long d = apx - (apy >> 1);
+        i32 d = apx - (apy >> 1);
         while ( true )
         {
           lbDisplay.GraphicsWindowPtr[offset] = ((unsigned char)colour);
@@ -139,7 +139,7 @@ static void draw_b_line(long x1, long y1, long x2, long y2, TbPixel colour)
       }
       else
       {
-        long d = apy - (apx >> 1);
+        i32 d = apy - (apx >> 1);
         while ( 1 )
         {
           lbDisplay.GraphicsWindowPtr[offset] = ((unsigned char)colour);
@@ -158,11 +158,11 @@ static void draw_b_line(long x1, long y1, long x2, long y2, TbPixel colour)
   }
 }
 
-TbResult LbDrawLine(long X1, long Y1, long X2, long Y2, TbPixel colour)
+TbResult LbDrawLine(i32 X1, i32 Y1, i32 X2, i32 Y2, TbPixel colour)
 {
     TbResult result = Lb_OK;
     // Adjusting X-dimension coordinates
-    long width_max = lbDisplay.GraphicsWindowWidth - 1;
+    i32 width_max = lbDisplay.GraphicsWindowWidth - 1;
     if ( X1 >= 0 )
     {
         if ( X1 <= width_max )
@@ -207,7 +207,7 @@ TbResult LbDrawLine(long X1, long Y1, long X2, long Y2, TbPixel colour)
         }
     }
     // Adjusting Y-dimension coordinates
-    long height_max = lbDisplay.GraphicsWindowHeight - 1;
+    i32 height_max = lbDisplay.GraphicsWindowHeight - 1;
     if ( Y1 < 0 )
     {
         if ( Y2 < 0 ) return 1;
