+++ bflibrary/src/general/poly_trigr_s.c
@@ -27,7 +27,7 @@
 # define EXPORT_SYMBOL(sym) STRINGIFY(sym)
 #endif
 
-extern const long add_to_edi[];
+extern const i32 add_to_edi[];
 
 /**
  * whether the subtraction (x-y) of two short ints would overflow
@@ -46,28 +46,28 @@ static inline ubyte __CFADDS__(short x, short y)
 }
 
 /**
- * whether the addition (x+y) of two long ints would use carry
+ * whether the addition (x+y) of two i32 ints would use carry
  */
-static inline ubyte __CFADDL__(long x, long y)
+static inline ubyte __CFADDL__(i32 x, i32 y)
 {
-    return (ulong)(x) > (ulong)(x+y);
+    return (u32)(x) > (u32)(x+y);
 }
 
 /**
- * rotate left unsigned long
+ * rotate left unsigned i32
  */
-static inline ulong __ROL4__(ulong value, int count)
+static inline u32 __ROL4__(u32 value, int count)
 {
     const uint nbits = 4 * 8;
 
     if (count > 0) {
         count %= nbits;
-        ulong high = value >> (nbits - count);
+        u32 high = value >> (nbits - count);
         value <<= count;
         value |= high;
     } else {
         count = -count % nbits;
-        ulong low = value << (nbits - count);
+        u32 low = value << (nbits - count);
         value >>= count;
         value |= low;
     }
